# src/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Cool Project Station",
    "description": "My User Project Application API"
    
  },
  "host": "https://project-cool-station.onrender.com",
  "basePath": "/",
  "tags": [
    {
      "name": "Cool Project Station",
      "description": "API CRUD para proyectos molones"
    }
  ],

  "paths": {
    "/api/projects/add": {
      "post": {
        "tags": ["Personal Project Card"],
        "description": "Crear un nuevo proyecto con toda su info",
        "parameters": [
          {
            "autor": "autor",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "job": "job",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "image": "image",
            "in": "body",
            "type": "string"
          },
          {
            "photo": "photo",
            "in": "body",
            "type": "string"
          },
          {
            "name": "name",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "slogan": "slogan",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "repo": "repo",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "demo": "demo",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "technologies": "technologies",
            "in": "body",
            "required": true,
            "type": "string"
          },
          {
            "description": "description",
            "in": "body",
            "required": true,
            "type": "string"
          }
        ],
        "produces": ["application/json"],
        "responses": {
          "200": {
            "success": "https://project-cool-station.onrender.com/api/projects/detail/${results.insertId}"
          
        }
      }
    }
  },
  "/api/projects/all": {
    "get": {
      "tags": ["projects"],
      "summary": "Get all data for the projects",
      "responses": {
        "200": {
          "description": "OK",
          "schema": {
          }
        }
      }
    }
  },
  "/api/projects/detail/:projectID": {
    "get": {
      "tags": ["projects detail"],
       "parameters": [
        {
          "name": "projectId",
          "in": "path",
          "required": true,
          "description": "ID of project that we want to find",
          "type": "integer"
        }
      ],
      "summary": "Get detail data for one project by projectID ",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              
            }
          }
        }
      }
    },
    "/api/projects/delete_all": {
      "delete": {
        "tags": ["projects"],
        "summary": "Delete all projects in database",
        "responses": {
          "200": {
            "description": "All projects are deleted",
            "schema": {
              
            }
          }
        }
      }
    },
    "/api/projects/delete_one/:idCard": {
      "delete": {
        "tags": ["projects"],
        "parameters": [
          {
            "name": "idCard",
            "in": "path",
            "required": true,
            "description": "ID of card that we want to find to delete",
            "type": "integer"
          }
        ],
        "summary": "Delete project with given ID",
        "responses": {
          "200": {
            "description": "Project is deleted",
            "schema": {
            }
          }
        }
      }
    }
  },

  "schemes": ["http"],
  "consumes": ["application/json"],
  "produces": ["application/json"]
}